% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\name{academic_colour_palette_names}
\alias{academic_colour_palette_names}
\title{Get The Academic Colour Palette Names}
\usage{
academic_colour_palette_names()
}
\value{
A vector of the names of the colour palettes available in \code{AcademicThemes}.
}
\description{
Get The Academic Colour Palette Names
}
\examples{
academic_colour_palette_names()
}
